Fixed all the tests in sPyNNakerExternalDevicesPlugin and SpiNNMan, except test_stop_receiving_items_from_queue.
Modified various function code and interfaces following unit tests results and requirements

Corrected a bug in the buffered injectr, for which a spike_source_array population could not be instantiated without any spike to inject.

Modified the transceiver to remove the "discover" option from the class instantiation, as this required the assumption that the board was already booted (which was not a strict requirement).