Fix critical bug in LPG caused by mixup between & and &&
Tests of a bit flag must always be done with & and not &&; the latter
has very different semantics indeed! To help prevent this sort of
problem in future, I've defined a few helpful simple macros that
describe how exactly to test a bit flag, and also how to clamp integers
to their low 8 and 16 bits (because these are common and areas that it
is easy to make the same sort of mistake in).