Turn warnings into errors when building C code in FEC.
In my experience, warnings are usually good canaries for bad code, and
it is worthwhile trying to ensure that code is free of warnings. It's
more achievable than usual in SpiNNaker as we use a clearly defined
toolchain; the locations of errors and their fixes are both systematic
and predictable.