No more layout enums
C has good ways to represent data structures in memory so that you
virtually never need to do complex address arithmetic (especially when
everything is word-aligned). This commit is about converting our code to
actually use those techniques.

The only one of our protocols that is resistant to this approach is
EIEIO, because that was written by someone truly twisted.