Move the timing data extraction to where it belongs.
The FinaliseTimingData algorithm is rather aware of what is going on
inside AbstractSpinnakerBase (as it knows that the active simulator is
one of those and what one of the non-public methods of the simulator is)
but it is better to have this one place be aware than hack things all
over.

This lets us use tokens to get the timing finalisation actually correct;
the writing of the provenance to local disk can be outside the timed
workflow zone.