Cheaper disable/enable of interrupts in logging
Because interrupt manipulation *cannot* be optimised out by the compiler
(as it doesn't really understand it) the placement of calls outside the
check of whether a logging call would be done meant that every log call,
even the disabled ones, resulted in two expensive calls being inserted.
We really don't want that!