Real bitfields in C
This uses real C bitfields (NB: not the ones that Alan is struggling
with! Not a close relation at all!) to make scrobbling around in the
SpiNNaker hardware nicer, which reduces the nastiness of the extra
monitor quite a bit. C bitfields are not particularly well known because
they're rather non-portable, but the ARM ABI locks them down more
(definitely little endian!) and as long as all the bitfield-structs are
word sized, use uints for all members, and cover all 32 bits, things
should be fine.

This does not make using the hardware *much* easier. You still need to
read the datasheet if you're going to do very much.