Prevent a buffer overrun in the LPG when using packets with payloads
The overrun was because of a miscalculation of the number of words to
clear after sending a message, yet it turns out that you don't need to
clear the memory at all.

Also factor out all the actual writes (not of just the unaligned words)
to helper inline functions to make the code easier to debug.

Fixes #593