Remove weird pattern that failed in production
The way to require an interface to be implemented by a subclass of some
class (e.g. MachineVertex) is to make it a subclass of that class. Tada!
Instant perfect enforcement (provided the slots are empty). If users get
it right, everything works. If they get it wrong, kaboom at definition
time, which we'll pick up during internal testing.