Tricky case!
Though the methods in AbstractSendMeMulticastCommandsVertex only apply
when the subclass is also a vertex, it's not actually guaranteed that
the subclasses ever are vertices at all; they could instead be proxied
on the machine by sending packets to host. The type system cannot really
cope with how we do this now; we have to relax the constraints.