A better error when the machine can't be talked to
Can't be more specific; we don't really know why the network comms
failed and there are many possible reasons (up to and including someone
drilling through a cable at the wrong moment!). But a down VPN is one of
the more likely options, and generating the real machine is definitely
when we actually first detect a problem.