That should make the transaction handling work correctly
Setting the internal isolation level to None means that we take care of
transactions ourselves *OR* that we let the SQLite core do what it wants
for us. Which will cover us for anywhere we've missed. What it does do
for sure is stop the Python standard library layer from getting in the
way; it has some funny ideas that don't work well with multiprocess use.